--- code/benchmark/internal/BenchmarkN.java
@@ -1,14 +1,0 @@
-package benchmark.internal;
-
-
-public class BenchmarkN {
-
-	public static void alloc(int id) {
-
-	}
-
-	public static void test(int id, Object targetVariable) {
-
-	}
-
-}
