--- code/benchmark/objects/B.java
@@ -1,9 +1,0 @@
-package benchmark.objects;
-
-public class B {
-
-	// Object B used as attribute of objects of type A
-
-	public B() {
-	}
-}
