--- code/benchmark/objects/I.java
@@ -1,7 +1,0 @@
-package benchmark.objects;
-
-public interface I {
-	// G and H implement I
-
-	public A foo(A a);
-}
