--- code/mytest/META-INF/MANIFEST.MF
@@ -1,3 +1,0 @@
-Manifest-Version: 1.0
-Created-By: 19 (Oracle Corporation)
-Main-Class: test.Fuck
