--- code/mytest/Var.java
@@ -1,5 +1,0 @@
-package test;
-
-public class Var{
-    public int i = 1;
-}
