+++ test_code/benchmark/internal/Benchmark.java
@@ -1,0 +1,18 @@
+package benchmark.internal;
+
+
+public class Benchmark {
+
+	public static void alloc(int id) {
+
+	}
+	/*
+	public static void test(int id, String targetVariable) {
+
+	}
+	*/	
+	public static void test(int id, Object targetVariable) {
+
+	}
+
+}
