+++ test_code/benchmark/objects/Son.java
@@ -1,0 +1,10 @@
+package benchmark.objects;
+
+public class Son extends Father {
+
+	public B pb;
+
+	public Son(A a) {
+		super(a);
+	}
+}
